🚚 separate endpoints in `docker_services`
