+++ backend/zane_api/tests/deployment.py
@@ -1498,7 +1498,7 @@ class DockerServiceDeploymentCancelChangesViewTests(AuthAPITestCase):
 
         response = self.client.delete(
             reverse(
-                "zane_api:services.cancel_deployment_changes",
+                "zane_api:services.cancel_service_changes",
                 kwargs={
                     "project_slug": p.slug,
                     "env_slug": "production",
@@ -1518,7 +1518,7 @@ class DockerServiceDeploymentCancelChangesViewTests(AuthAPITestCase):
 
         response = self.client.delete(
             reverse(
-                "zane_api:services.cancel_deployment_changes",
+                "zane_api:services.cancel_service_changes",
                 kwargs={
                     "project_slug": p.slug,
                     "env_slug": "production",
@@ -1553,7 +1553,7 @@ class DockerServiceDeploymentCancelChangesViewTests(AuthAPITestCase):
 
         response = self.client.delete(
             reverse(
-                "zane_api:services.cancel_deployment_changes",
+                "zane_api:services.cancel_service_changes",
                 kwargs={
                     "project_slug": p.slug,
                     "env_slug": "production",
