+++ backend/zane_api/views/logs.py
@@ -2,7 +2,7 @@ import json
 from urllib.parse import urlparse
 
 from drf_spectacular.utils import extend_schema
-from rest_framework import status
+from rest_framework import status, exceptions
 from rest_framework.request import Request
 from rest_framework.response import Response
 from rest_framework.throttling import ScopedRateThrottle
@@ -18,11 +18,39 @@ from .serializers import (
     DockerContainerLogsRequestSerializer,
     HTTPServiceLogSerializer,
 )
-from ..models import HttpLog
 from search.dtos import RuntimeLogDto, RuntimeLogLevel, RuntimeLogSource
 from search.loki_client import LokiSearchClient
 from django.conf import settings
 from django.utils import timezone
+from typing import cast
+
+from django_filters.rest_framework import DjangoFilterBackend
+from rest_framework.generics import ListAPIView, RetrieveAPIView
+
+
+from search.serializers import RuntimeLogsSearchSerializer
+
+from .base import EMPTY_CURSOR_RESPONSE
+
+from .serializers import (
+    DeploymentBuildLogsQuerySerializer,
+    DeploymentRuntimeLogsQuerySerializer,
+    HttpLogFieldsQuerySerializer,
+    HttpLogFieldsResponseSerializer,
+    DeploymentHttpLogsPagination,
+    DeploymentHttpLogsFilterSet,
+)
+
+from ..models import (
+    Project,
+    Service,
+    Deployment,
+    HttpLog,
+    Environment,
+)
+from ..serializers import HttpLogSerializer
+
+from rest_framework.utils.serializer_helpers import ReturnDict
 
 
 @extend_schema(exclude=True)
@@ -175,3 +203,439 @@ class LogIngestAPIView(APIView):
             )
             print(f"HTTP logs inserted = {Colors.BLUE}{len(http_logs)}{Colors.ENDC}")
             return Response(response.data, status=status.HTTP_200_OK)
+
+
+class ServiceHttpLogsFieldsAPIView(APIView):
+    serializer_class = HttpLogFieldsResponseSerializer
+
+    @extend_schema(
+        summary="Get service http logs fields values",
+        parameters=[HttpLogFieldsQuerySerializer],
+    )
+    def get(
+        self,
+        request: Request,
+        project_slug: str,
+        service_slug: str,
+        env_slug: str = Environment.PRODUCTION_ENV,
+    ):
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+            environment = Environment.objects.get(
+                name=env_slug.lower(), project=project
+            )
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+        else:
+            form = HttpLogFieldsQuerySerializer(data=request.query_params)
+            if form.is_valid(raise_exception=True):
+                field = form.data["field"]  # type: ignore
+                value = form.data["value"]  # type: ignore
+
+                condition = {}
+                if len(value) > 0:
+                    condition = {f"{field}__startswith": value}
+
+                values = (
+                    HttpLog.objects.filter(
+                        service_id=service.id,
+                        **condition,
+                    )
+                    .order_by(field)
+                    .values_list(field, flat=True)
+                    .distinct()[:7]
+                )
+
+                seriaziler = HttpLogFieldsResponseSerializer([item for item in values])
+                return Response(seriaziler.data)
+
+
+class ServiceHttpLogsAPIView(ListAPIView):
+    serializer_class = HttpLogSerializer
+    queryset = (
+        HttpLog.objects.all()
+    )  # This is to document API endpoints with drf-spectacular, in practive what is used is `get_queryset`
+    pagination_class = DeploymentHttpLogsPagination
+    filter_backends = [DjangoFilterBackend]
+    filterset_class = DeploymentHttpLogsFilterSet
+
+    @extend_schema(
+        summary="Get service HTTP logs",
+    )
+    def get(self, request: Request, *args, **kwargs):
+        try:
+            print("====== HTTP LOGS SEARCH ======")
+            print(f"Params: {Colors.GREY}{request.query_params}{Colors.ENDC}")
+            return super().get(request, *args, **kwargs)
+        except exceptions.NotFound as e:
+            if "Invalid cursor" in str(e.detail):
+                return Response(EMPTY_CURSOR_RESPONSE)
+            raise e
+
+    def get_queryset(self):  # type: ignore
+        project_slug = self.kwargs["project_slug"]
+        service_slug = self.kwargs["service_slug"]
+        env_slug = self.kwargs.get("env_slug") or Environment.PRODUCTION_ENV
+
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+
+            environment = Environment.objects.get(
+                name=env_slug.lower(), project=project
+            )
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            return service.http_logs
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+
+
+class ServiceSingleHttpLogAPIView(RetrieveAPIView):
+    serializer_class = HttpLogSerializer
+    queryset = (
+        HttpLog.objects.all()
+    )  # This is to document API endpoints with drf-spectacular, in practive what is used is `get_queryset`
+    lookup_url_kwarg = "request_uuid"  # This corresponds to the URL configuration
+
+    @extend_schema(summary="Get single service http log")
+    def get(self, request, *args, **kwargs):
+        return super().get(request, *args, **kwargs)
+
+    def get_object(self):  # type: ignore
+        project_slug = self.kwargs["project_slug"]
+        service_slug = self.kwargs["service_slug"]
+        request_uuid = self.kwargs["request_uuid"]
+        env_slug = self.kwargs.get("env_slug") or Environment.PRODUCTION_ENV
+
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+
+            environment = Environment.objects.get(name=env_slug, project=project)
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            http_log = service.http_logs.filter(
+                service_id=service.id, request_id=request_uuid
+            ).first()
+
+            if http_log is None:
+                raise exceptions.NotFound(
+                    detail=f"A HTTP log with the id of `{request_uuid}` does not exist for this deployment."
+                )
+            return http_log
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+
+
+class ServiceDeploymentRuntimeLogsAPIView(APIView):
+    serializer_class = RuntimeLogsSearchSerializer
+
+    @extend_schema(
+        summary="Get deployment logs", parameters=[DeploymentRuntimeLogsQuerySerializer]
+    )
+    def get(
+        self,
+        request: Request,
+        project_slug: str,
+        service_slug: str,
+        deployment_hash: str,
+        env_slug: str = Environment.PRODUCTION_ENV,
+    ):
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+
+            environment = Environment.objects.get(
+                name=env_slug.lower(), project=project
+            )
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            deployment = Deployment.objects.get(service=service, hash=deployment_hash)
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+        except Deployment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A deployment with the hash `{deployment_hash}` does not exist for this service."
+            )
+        else:
+            form = DeploymentRuntimeLogsQuerySerializer(data=request.query_params)
+            print(f"{request.query_params=}")
+            if form.is_valid(raise_exception=True):
+                search_client = LokiSearchClient(host=settings.LOKI_HOST)
+                data = search_client.search(
+                    query=dict(**form.validated_data, deployment_id=deployment.hash),  # type: ignore
+                )
+                return Response(data)
+
+
+class ServiceDeploymentBuildLogsAPIView(APIView):
+    serializer_class = RuntimeLogsSearchSerializer
+
+    @extend_schema(
+        summary="Get deployment build logs",
+        parameters=[DeploymentBuildLogsQuerySerializer],
+    )
+    def get(
+        self,
+        request: Request,
+        project_slug: str,
+        service_slug: str,
+        deployment_hash: str,
+        env_slug: str = Environment.PRODUCTION_ENV,
+    ):
+        try:
+            project = Project.objects.get(slug=project_slug)
+
+            environment = Environment.objects.get(
+                name=env_slug.lower(), project=project
+            )
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            deployment = Deployment.objects.get(service=service, hash=deployment_hash)
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+        except Deployment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A deployment with the hash `{deployment_hash}` does not exist for this service."
+            )
+        else:
+            form = DeploymentBuildLogsQuerySerializer(data=request.query_params)
+            print(f"{request.query_params=}")
+            if form.is_valid(raise_exception=True):
+                search_client = LokiSearchClient(host=settings.LOKI_HOST)
+                data = search_client.search(
+                    query=dict(
+                        cursor=cast(ReturnDict, form.validated_data).get("cursor"),
+                        deployment_id=deployment.hash,
+                        source=[RuntimeLogSource.BUILD, RuntimeLogSource.SYSTEM],
+                    ),  # type: ignore
+                )
+                return Response(data)
+
+
+class ServiceDeploymentHttpLogsFieldsAPIView(APIView):
+    serializer_class = HttpLogFieldsResponseSerializer
+
+    @extend_schema(
+        summary="Get deployment http logs fields values",
+        parameters=[HttpLogFieldsQuerySerializer],
+    )
+    def get(
+        self,
+        request: Request,
+        project_slug: str,
+        service_slug: str,
+        deployment_hash: str,
+        env_slug: str = Environment.PRODUCTION_ENV,
+    ):
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+
+            environment = Environment.objects.get(
+                name=env_slug.lower(), project=project
+            )
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            deployment = Deployment.objects.get(service=service, hash=deployment_hash)
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+        except Deployment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A deployment with the hash `{deployment_hash}` does not exist for this service."
+            )
+        else:
+            form = HttpLogFieldsQuerySerializer(data=request.query_params)
+            if form.is_valid(raise_exception=True):
+                field = form.data["field"]  # type: ignore # type: ignore
+                value = form.data["value"]  # type: ignore # type: ignore
+
+                condition = {}
+                if len(value) > 0:
+                    condition = {f"{field}__startswith": value}
+
+                values = (
+                    HttpLog.objects.filter(
+                        deployment_id=deployment.hash,
+                        service_id=service.id,
+                        **condition,
+                    )
+                    .order_by(field)
+                    .values_list(field, flat=True)
+                    .distinct()[:7]
+                )
+
+                seriaziler = HttpLogFieldsResponseSerializer([item for item in values])
+                return Response(seriaziler.data)
+
+
+class ServiceDeploymentHttpLogsAPIView(ListAPIView):
+    serializer_class = HttpLogSerializer
+    queryset = (
+        HttpLog.objects.all()
+    )  # This is to document API endpoints with drf-spectacular, in practive what is used is `get_queryset`
+    pagination_class = DeploymentHttpLogsPagination
+    filter_backends = [DjangoFilterBackend]
+    filterset_class = DeploymentHttpLogsFilterSet
+
+    @extend_schema(
+        summary="Get deployment HTTP logs",
+    )
+    def get(self, request, *args, **kwargs):
+        try:
+            return super().get(request, *args, **kwargs)
+        except exceptions.NotFound as e:
+            if "Invalid cursor" in str(e.detail):
+                return Response(EMPTY_CURSOR_RESPONSE)
+            raise e
+
+    def get_queryset(self):  # type: ignore
+        project_slug = self.kwargs["project_slug"]
+        service_slug = self.kwargs["service_slug"]
+        deployment_hash = self.kwargs["deployment_hash"]
+        env_slug = self.kwargs.get("env_slug") or Environment.PRODUCTION_ENV
+
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+            environment = Environment.objects.get(
+                name=env_slug.lower(), project=project
+            )
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            deployment = Deployment.objects.get(service=service, hash=deployment_hash)
+            return deployment.http_logs
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+        except Deployment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A deployment with the hash `{deployment_hash}` does not exist for this service."
+            )
+
+
+class ServiceDeploymentSingleHttpLogAPIView(RetrieveAPIView):
+    serializer_class = HttpLogSerializer
+    queryset = (
+        HttpLog.objects.all()
+    )  # This is to document API endpoints with drf-spectacular, in practive what is used is `get_queryset`
+    lookup_url_kwarg = "request_uuid"  # This corresponds to the URL configuration
+
+    @extend_schema(summary="Get single deployment http log")
+    def get(self, request, *args, **kwargs):
+        return super().get(request, *args, **kwargs)
+
+    def get_object(self):  # type: ignore
+        project_slug = self.kwargs["project_slug"]
+        service_slug = self.kwargs["service_slug"]
+        deployment_hash = self.kwargs["deployment_hash"]
+        request_uuid = self.kwargs["request_uuid"]
+        env_slug = self.kwargs.get("env_slug") or Environment.PRODUCTION_ENV
+
+        try:
+            project = Project.objects.get(slug=project_slug, owner=self.request.user)
+
+            environment = Environment.objects.get(name=env_slug, project=project)
+            service = Service.objects.get(
+                slug=service_slug, project=project, environment=environment
+            )
+            deployment = Deployment.objects.get(service=service, hash=deployment_hash)
+            http_log = deployment.http_logs.filter(
+                deployment_id=deployment_hash, request_id=request_uuid
+            ).first()
+
+            if http_log is None:
+                raise exceptions.NotFound(
+                    detail=f"A HTTP log with the id of `{request_uuid}` does not exist for this deployment."
+                )
+            return http_log
+        except Project.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A project with the slug `{project_slug}` does not exist."
+            )
+        except Environment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"An environment with the name `{env_slug}` does not exist in this project"
+            )
+        except Service.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A service with the slug `{service_slug}` does not exist within the environment `{env_slug}` of the project `{project_slug}`"
+            )
+        except Deployment.DoesNotExist:
+            raise exceptions.NotFound(
+                detail=f"A deployment with the hash `{deployment_hash}` does not exist for this service."
+            )
